POND-19: add new types, fix class prop, add translations
