+++ types/cms/CmsImageSlider/CmsImageSliderItem.d.ts
@@ -1,0 +1,7 @@
+export type CmsImageSliderItem = {
+  url: string;
+  newTab: boolean;
+  media: Schemas['Media'];
+  mediaId: string;
+  apiAlias: 'cms_image_slider_item';
+};
