+++ types/cms/CmsImageSlider/NavigationOptions.d.ts
@@ -1,0 +1,1 @@
+export type NavigationOptions = 'none' | 'inside' | 'outside';
