+++ types/cms/CmsImageSlider/SwiperBreakpointOptions.d.ts
@@ -1,0 +1,6 @@
+export type SliderBreakpointOptions = {
+  [width: number]: {
+    slidesPerView?: number;
+    spaceBetween?: number;
+  };
+};
